refactor(runtime): extract SequentialDispatcher from HeadlessDelegate
W6.2a (ADR-160 L2 prep). Pure refactor — byte-identical behavior to the
original `HeadlessDelegate::execute_tool_calls` 5-step pipeline
(ToolCallStart -> await_approval_for -> pre EgressGate -> executor.execute
-> post EgressGate + sanitizer + ToolResult).

Moves the 173-line inline loop body into a new `pub(crate) struct
SequentialDispatcher` in `tool_dispatch.rs`, which owns Arc handles to
the executor, egress gate, approval policy, approval inbox, sanitizer and
event channel. `HeadlessDelegate::execute_tool_calls` now constructs a
dispatcher and delegates.

Sets the stage for W6.2b (introduce `ToolDispatcher` trait and expose
`SequentialDispatcher` as the default impl) without yet changing the
public surface.

Verification:
- cargo check -p dasclaw_runtime --tests (0 errors, 0 warnings)
- cargo nextest run -p dasclaw_runtime (196/196 passed)
- cargo clippy --no-deps -p dasclaw_runtime --all-targets -- -D warnings (clean)
- cargo fmt --all
- python3.12 scripts/check_no_panics.py --base origin/xClaw (OK)

Cross-cuts: 类A (no .ironclaw / IRONCLAW_BASE_DIR literals added)

Refs ADR-160 §3 (5-layer model, L2 ToolDispatcher seam)
